requirements.txt
streamlit==1.25.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
scikit-learn==1.3.0
joblib==1.3.1
openpyxl==3.1.2
pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
